Implemented feed view in the application
