+++ client/src/sections/feed/view/feed-view.jsx
@@ -2,14 +2,14 @@ import { useState } from 'react';
 
 import Stack from '@mui/material/Stack';
 import Container from '@mui/material/Container';
-import Grid from '@mui/material/Unstable_Grid2';
 import Typography from '@mui/material/Typography';
 
-import { products } from 'src/_mock/products';
+// eslint-disable-next-line import/no-unresolved
+import { repositories } from 'src/_mock/repositories';
 
-import ProductCard from '../product-card';
-import ProductSort from '../product-sort';
-import ProductFilters from '../product-filters';
+import ProductSort from '../sort';
+import FeedCards from '../feed-cards';
+import ProductFilters from '../filters';
 
 // ----------------------------------------------------------------------
 
@@ -48,13 +48,9 @@ export default function ProductsView() {
         </Stack>
       </Stack>
 
-      <Grid container spacing={3}>
-        {products.map((product) => (
-          <Grid key={product.id} xs={12} sm={6} md={3}>
-            <ProductCard product={product} />
-          </Grid>
-        ))}
-      </Grid>
+      <Stack>
+        <FeedCards items={repositories} />
+      </Stack>
     </Container>
   );
 }
