+++ client/src/sections/feed/view/index.js
@@ -1,1 +1,1 @@
-export { default as ProductsView } from './products-view';
+export { default as ProductsView } from './feed-view';
